[delete-user-fix] Fix bug that deletes user when you delete the default user character
And also add a confirmation modal when deleting a user character

+++ client/app/modules/user-management/components/user-character-row.component.ts
@@ -74,7 +74,16 @@ export class UserCharacterRowComponent implements OnInit {
     );
   }
   public deleteUserCharacter() {
-    this.userService.deleteUserCharacter(this.userCharacter);
+    this.commonUxService.openConfirmModal(
+      `Remove saved character ${this.userCharacter.characterName}?`,
+      'Delete saved character',
+      false,
+      'Delete it').then(
+        confirm => {
+          this.userService.deleteUserCharacter(this.userCharacter);
+        }, dismiss => {
+          // Cancelled. Do nothign
+        });
   }
   public setDefaultUserCharacter() {
     this.userService.setDefaultUserCharacter(this.userCharacter);
